Random Password Generator
This program will create a randomly generated password. All you have to do is input your preferred length, however the limit for your inputted length for your randomized password is 70. So if you input a length of 71 and higher, you will not generate a password and it will respond with the following message: "Password length too long."